fix(web): gate ProjectPicker's cloud-daemon offer on compute eligibility
ProjectPicker's ConnectDaemonModal offered "Reliant Cloud — Provision a hosted
daemon now" gated only on `capabilities.cloudDaemons`. That flag answers "does
this DEPLOYMENT have cloud daemons", which is not the question "may THIS
ACCOUNT start one". Eligibility was consulted nowhere in this file.

So an unfunded user got a fully-enabled button, and clicking it really did
reach CreateDaemon. A brand-new production user hit exactly this after being
ejected out of onboarding into the picker.

Onboarding's ComputeStep has always gated the same action on
useCloudEligibility() and shows the reason plus the coupon/plans affordances
when the answer is no. This modal is the other door into that action, so it now
asks the same question and offers the same way out. Eligibility that is still
loading counts as not-yet-permitted rather than allowed.

The modal moves to its own file so the gate is directly testable; ProjectPicker
was past 1700 lines and the modal was private to it. No behavior change for the
self-hosted path or for an eligible user.

The server remains the authority — the server-side gap in internal/svcdaemon is
tracked separately and is NOT addressed here. This only stops the client
committing to a flow it has already been told will be refused.

Verified by reverting the gate: "does NOT provision when ineligible" and the
still-loading case both fail without it (CreateDaemon fires), pass with it.
Full web suite 1903 passed / 255 files.
