+++ web/src/components/Projects/ProjectPicker.tsx
@@ -5,9 +5,7 @@ import {
   GitFork,
   Cloud,
   Loader2,
-  Check,
   ArrowLeft,
-  Monitor,
   Search,
   X,
   Pencil,
@@ -30,8 +28,8 @@ import { RepoSelector } from "./RepoSelector";
 
 import { toast } from "../../lib/toast-manager";
 import { useDaemonStatus } from "../../hooks/useDaemonStatus";
-import { useResumeDaemon, useCreateDaemon } from "../../hooks/useOnboardingQueries";
-import { SelfHostedDaemonConnect } from "./SelfHostedDaemonConnect";
+import { useResumeDaemon } from "../../hooks/useOnboardingQueries";
+import { ConnectDaemonModal } from "./ConnectDaemonModal";
 // Aliased deliberately. There are TWO unrelated `DaemonStatus` enums in this
 // app and their numeric values COLLIDE: registry ACTIVE=1 is control-plane
 // PENDING=1, registry IDLE=2 is control-plane ACTIVE=2. This file reads both
@@ -221,220 +219,6 @@ function SortableHeader({
         )}
       </button>
     </th>
-  );
-}
-
-const MANAGED_DAEMON_TYPE = 1;
-const MANAGED_DAEMON_SIZE_SMALL = 1;
-
-// ConnectDaemonModal — the in-place "Connect a new daemon" flow. Offers BOTH
-// connect paths without bouncing the user into the onboarding wizard:
-//   - Managed cloud daemon: provisions a hosted daemon via CreateDaemon (only
-//     offered when this deployment has cloud daemons).
-//   - Self-hosted daemon: the SelfHostedDaemonConnect panel (shared with
-//     onboarding's ComputeStep) — generates a token + shows the
-//     `reliant daemon start --token` install steps.
-// The modal auto-dismisses once a daemon connects (the picker rerenders into
-// its normal project list as soon as useDaemonStatus reports an active daemon).
-function ConnectDaemonModal({
-  isOpen,
-  onClose,
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-}) {
-  const hasCloud = capabilities.cloudDaemons;
-  // "cloud" | "self" | null — null shows the choice, the others show the
-  // respective flow. Default to the choice screen so the user explicitly
-  // picks how they want to connect.
-  const [mode, setMode] = useState<"cloud" | "self" | null>(null);
-  const [cloudError, setCloudError] = useState<string | null>(null);
-  const [cloudStarted, setCloudStarted] = useState(false);
-
-  const createDaemonMutation = useCreateDaemon({
-    onError: (err) => {
-      const msg = err instanceof Error ? err.message : "Failed to start cloud daemon";
-      setCloudError(msg);
-    },
-  });
-
-  // Reset to the choice screen each time the modal opens so a prior session's
-  // mode doesn't leak in.
-  useEffect(() => {
-    if (isOpen) {
-      setMode(null);
-      setCloudError(null);
-      setCloudStarted(false);
-    }
-  }, [isOpen]);
-
-  const startCloudDaemon = useCallback(async () => {
-    setCloudError(null);
-    try {
-      const { listDaemons, hasActiveDaemon } = await import(
-        "../../services/controlPlane/daemon"
-      );
-      const { daemons } = await listDaemons();
-      if (hasActiveDaemon(daemons)) {
-        // Already have an active daemon — nothing to provision; the picker
-        // will pick it up on the next poll.
-        setCloudStarted(true);
-        return;
-      }
-      if (daemons.length > 0) {
-        // A suspended/disconnected daemon exists — resume it instead of
-        // creating a duplicate. Resume failures are non-fatal; the user can
-        // retry from the "Resume a daemon" list.
-        const id = daemons[0]?.id ?? "";
-        if (id) {
-          const { resumeDaemon } = await import(
-            "../../services/controlPlane/daemon"
-          );
-          await resumeDaemon(id);
-          setCloudStarted(true);
-          return;
-        }
-      }
-      await createDaemonMutation.mutateAsync({
-        name: "workspace-daemon",
-        daemonType: MANAGED_DAEMON_TYPE,
-        size: MANAGED_DAEMON_SIZE_SMALL,
-        gitRepo: "",
-        gitBranch: "main",
-      });
-      setCloudStarted(true);
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : "Failed to start cloud daemon";
-      setCloudError(msg);
-    }
-  }, [createDaemonMutation]);
-
-  const startingCloud = createDaemonMutation.isPending;
-
-  return (
-    <Modal isOpen={isOpen} onClose={onClose} title="Connect a daemon" size="lg">
-      {mode === null && (
-        <div className="space-y-3">
-          <p className="text-sm text-muted-foreground">
-            A daemon runs your code. Connect one to start working — either a
-            hosted Reliant Cloud daemon or your own self-hosted machine.
-          </p>
-          <div className="grid gap-3 sm:grid-cols-2">
-            <button
-              type="button"
-              onClick={() => {
-                setMode("cloud");
-                void startCloudDaemon();
-              }}
-              disabled={!hasCloud}
-              className={`flex min-w-0 flex-col items-start gap-3 rounded-xl border-2 p-5 text-left transition-all ${
-                hasCloud
-                  ? "border-primary/25 bg-primary/5 hover:border-primary/50 hover:bg-primary/10"
-                  : "cursor-not-allowed border-border/50 bg-muted/30 opacity-70"
-              }`}
-            >
-              <div className="rounded-lg bg-primary/15 p-2.5 text-primary">
-                <Cloud className="h-6 w-6" />
-              </div>
-              <div className="space-y-1">
-                <span className="block text-sm font-semibold text-foreground">
-                  Reliant Cloud
-                </span>
-                <span className="block text-xs leading-relaxed text-muted-foreground">
-                  {hasCloud
-                    ? "Provision a hosted daemon now. No install required."
-                    : "Cloud daemons are not enabled for this deployment."}
-                </span>
-              </div>
-            </button>
-
-            <button
-              type="button"
-              onClick={() => setMode("self")}
-              className="flex min-w-0 flex-col items-start gap-3 rounded-xl border-2 border-border/50 bg-background p-5 text-left transition-all hover:border-primary/50 hover:bg-muted/50"
-            >
-              <div className="rounded-lg bg-muted p-2.5 text-muted-foreground">
-                <Monitor className="h-6 w-6" />
-              </div>
-              <div className="space-y-1">
-                <span className="block text-sm font-semibold text-foreground">
-                  Self-hosted
-                </span>
-                <span className="block text-xs leading-relaxed text-muted-foreground">
-                  Run the daemon on your own laptop or server with a token.
-                </span>
-              </div>
-            </button>
-          </div>
-        </div>
-      )}
-
-      {mode === "cloud" && (
-        <div className="space-y-4">
-          <button
-            type="button"
-            onClick={() => setMode(null)}
-            className="inline-flex items-center gap-1.5 text-xs text-muted-foreground transition-colors hover:text-foreground"
-          >
-            <ArrowLeft className="h-3.5 w-3.5" />
-            Back
-          </button>
-          <div className="flex items-center gap-3 rounded-xl border border-border/50 bg-muted/30 p-4">
-            {startingCloud ? (
-              <Loader2 className="h-5 w-5 animate-spin text-primary" />
-            ) : cloudStarted ? (
-              <Check className="h-5 w-5 text-emerald-500" />
-            ) : (
-              <Cloud className="h-5 w-5 text-primary" />
-            )}
-            <div className="min-w-0">
-              <h3 className="text-sm font-medium text-foreground">
-                {startingCloud
-                  ? "Requesting a cloud daemon..."
-                  : cloudStarted
-                    ? "Cloud daemon requested"
-                    : "Reliant Cloud"}
-              </h3>
-              <p className="mt-0.5 text-xs text-muted-foreground">
-                {cloudStarted
-                  ? "It may take a few minutes to provision. This screen will refresh once it connects."
-                  : "Provisioning a hosted daemon for your account."}
-              </p>
-            </div>
-          </div>
-          {cloudError && (
-            <div className="space-y-2">
-              <p className="text-xs text-destructive">{cloudError}</p>
-              <button
-                type="button"
-                onClick={() => void startCloudDaemon()}
-                disabled={startingCloud}
-                className="w-full rounded-lg bg-primary px-4 py-2.5 text-sm font-semibold text-primary-foreground transition-colors hover:bg-primary/90 disabled:opacity-60"
-              >
-                Try again
-              </button>
-            </div>
-          )}
-        </div>
-      )}
-
-      {mode === "self" && (
-        <div className="space-y-4">
-          <button
-            type="button"
-            onClick={() => setMode(null)}
-            className="inline-flex items-center gap-1.5 text-xs text-muted-foreground transition-colors hover:text-foreground"
-          >
-            <ArrowLeft className="h-3.5 w-3.5" />
-            Back
-          </button>
-          {/* Shared with onboarding's ComputeStep — generates a token and
-              shows `reliant daemon start --token` install steps. Closes the
-              modal the moment the daemon connects. */}
-          <SelfHostedDaemonConnect onConnected={onClose} />
-        </div>
-      )}
-    </Modal>
   );
 }
